requirements.txt
openai
openai-whisper
sounddevice
soundfile

# TTS
PyQt6